# apps/root/src/app/api/comunicados/api-docs-comunicados.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Comunicados API",
        "description": "Documentação navegável dos contratos HTTP do módulo de Comunicados.",
        "version": "0.0.1-SNAPSHOT"
    },
    "servers": [
        {
            "url": "http://localhost:8085",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/posts/{postId}/images/presign": {
            "post": {
                "tags": [
                    "Imagens"
                ],
                "summary": "Solicitar URL pré-assinada para upload direto ao S3",
                "operationId": "presign",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PresignUploadRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "URL gerada — faça PUT direto ao S3",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PresignUploadResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "ARQ02 — limite de 5 arquivos atingido, tipo inválido ou tamanho excedido",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PresignUploadResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão de edição",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PresignUploadResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Comunicado não encontrado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PresignUploadResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/posts/publish": {
            "post": {
                "tags": [
                    "Postagens"
                ],
                "summary": "Publicar comunicado",
                "description": "Cria e publica o comunicado numa única transação (#107): nasce PUBLISHED, com destinos e histórico (CREATION + PUBLICATION). O autor/publicador é o usuário autenticado.",
                "operationId": "publish",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PublishAnnouncementRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Comunicado criado e publicado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnnouncementResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados obrigatórios ausentes ou inválidos",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnnouncementResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnnouncementResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão / fora do escopo (PA02/PA03)",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnnouncementResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{postId}/images/{imageId}/thumbnail": {
            "patch": {
                "tags": [
                    "Imagens"
                ],
                "summary": "Definir miniatura do comunicado",
                "operationId": "setThumbnail",
                "parameters": [
                    {
                        "name": "postId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "imageId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Miniatura definida",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnnouncementFileResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Sem permissão de edição",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnnouncementFileResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Arquivo não encontrado no comunicado",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/AnnouncementFileResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AnnouncementResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string",
                        "description": "Título do comunicado.",
                        "maxLength": 255
                    },
                    "description": {
                        "type": "string",
                        "description": "Conteúdo do comunicado em HTML sanitizado (TipTap)."
                    },
                    "descriptionPlain": {
                        "type": "string",
                        "description": "Versão plain-text da descrição, sem HTML — gerada no servidor."
                    },
                    "origin": {
                        "type": "string",
                        "enum": [
                            "WEG",
                            "SENAI",
                            "BOTH"
                        ]
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "SCHEDULED",
                            "PUBLISHED",
                            "REMOVED"
                        ]
                    },
                    "pinned": {
                        "type": "boolean"
                    },
                    "pinnedOrder": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "createdByUserId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "publishedByUserId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "scheduledFor": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "publishedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "removedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "AnnouncementFileResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "announcementId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "originalName": {
                        "type": "string"
                    },
                    "s3Key": {
                        "type": "string"
                    },
                    "s3Bucket": {
                        "type": "string"
                    },
                    "processedS3Key": {
                        "type": "string"
                    },
                    "displayUrl": {
                        "type": "string"
                    },
                    "contentType": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "IMAGE",
                            "DOCUMENT",
                            "VIDEO"
                        ]
                    },
                    "fileStatus": {
                        "type": "string",
                        "enum": [
                            "PENDING",
                            "READY",
                            "FAILED"
                        ]
                    },
                    "sizeBytes": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "isThumbnail": {
                        "type": "boolean"
                    },
                    "uploadedByUserId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PresignUploadRequest": {
                "type": "object",
                "properties": {
                    "contentType": {
                        "type": "string",
                        "minLength": 1
                    },
                    "sizeBytes": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "originalName": {
                        "type": "string",
                        "minLength": 1
                    },
                    "thumbnail": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "contentType",
                    "originalName"
                ]
            },
            "PresignUploadResponse": {
                "type": "object",
                "properties": {
                    "fileId": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "uploadUrl": {
                        "type": "string"
                    },
                    "fields": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "CreateAnnouncementDestinationInput": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "GENERAL",
                            "COURSE",
                            "CLASS",
                            "USER"
                        ]
                    },
                    "referenceId": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "type"
                ]
            },
...
            "PublishAnnouncementRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "description": "Título do comunicado.",
                        "example": "Reunião de alinhamento",
                        "maxLength": 255,
                        "minLength": 0
                    },
                    "description": {
                        "type": "string",
                        "description": "Conteúdo do comunicado em HTML sanitizado (allowlist TipTap: p, br, strong/b, em/i, u, ul/ol/li, a[href]). O servidor strip tags perigosas e deriva descriptionPlain automaticamente.",
                        "minLength": 1
                    },
                    "origin": {
                        "type": "string",
                        "enum": [
                            "WEG",
                            "SENAI",
                            "BOTH"
                        ]
                    },
                    "destinations": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CreateAnnouncementDestinationInput"
                        },
                        "minItems": 1
                    },
                    "pinned": {
                        "type": "boolean"
                    },
                    "tagIds": {
                        "type": "array",
                        "description": "UUIDs internos da tabela tag (tag.id), não hub_entity_id do Core. Opcional.",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    "shiftCodes": {
                        "type": "array",
                        "description": "Turnos do comunicado (enum Shift do Core). Opcional; vazio = sem restrição de turno.",
                        "example": [
                            "FULL_AM_PM"
                        ],
                        "items": {
                            "type": "string",
                            "enum": [
                                "FULL_AM_PM",
                                "FULL_PM_NT"
                            ]
                        }
                    },
                    "roles": {
                        "type": "array",
                        "description": "Papéis de usuário que podem ver o comunicado. Opcional; vazio = sem restrição de papel.",
                        "example": [
                            "STUDENT"
                        ],
                        "items": {
                            "type": "string",
                            "enum": [
                                "STUDENT",
                                "REPRESENTATIVE",
                                "TEACHER",
                                "SENAI",
                                "WEG",
                                "ADMIN"
                            ]
                        }
                    }
                },
                "required": [
                    "description",
                    "destinations",
                    "origin",
                    "title"
                ]
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}